# document/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PnP Web Project Backend API",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "https://SERVER_URL/api"
    }
  ],
  "paths": {
...
    "/search/book": {
      "description": "Search Book",
      "get": {
        "operationId": "SearchBook",
        "parameters": [
          {
            "$ref": "#/components/parameters/QSTitle"
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": {
                      "$ref": "#/components/schemas/Info"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": [
                            {
                              "$ref": "#/components/schemas/Book"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search/movie": {
      "description": "Search Movie",
      "get": {
        "operationId": "SearchMovie",
        "parameters": [
          {
            "$ref": "#/components/parameters/QSTitle"
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": {
                      "$ref": "#/components/schemas/Info"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": [
                            {
                              "$ref": "#/components/schemas/Movie"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "QSTitle": {
        "name": "title",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Info": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Book": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "publishedDate": {
            "type": "integer"
          },
          "thumbnail": {
            "type": "string"
          }
        }
      },
...
      "Movie": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "publishedYear": {
            "type": "integer"
          },
          "director": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}